move files to separate folders
