+++ level_progress/unlocks_test.go
@@ -1,4 +1,4 @@
-package wanikani
+package level_progress
 
 import (
 	"encoding/json"
